requirements.txt
python-telegram-bot==20.7
yt-dlp==2024.1.4
python-dotenv==1.0.0
aiohttp==3.9.1
